requirements.txt
alembic==1.3.2
atomicwrites==1.3.0
attrs==19.3.0
awesome-slugify==1.6.5
bcrypt==3.1.7
bleach==3.1.0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudinary==1.19.1
colorama==0.4.3
coverage==5.0.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.5.2
Flask-Minify==0.25
Flask-MySQL==1.5.0
Flask-Share==0.1.1
Flask-Sijax==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
Flask-Turbolinks==0.2.0
Flask-WTF==0.14.2
future==0.18.2
htmlmin==0.1.12
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsmin==2.2.2
lesscpy==0.14.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==8.0.2
mysql-connector-python==8.0.18
packaging==19.2
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
protobuf==3.11.2
pur==5.3.0
py==1.8.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
regex==2019.12.20
requests==2.22.0
Sijax==0.3.2
six==1.13.0
SQLAlchemy==1.3.12
twilio==6.35.2
Unidecode==1.1.1
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
wtf-tinymce==0.1.3
WTForms==2.2.1
xxhash==1.4.4
zipp==0.6.0
